Add files via upload
This project demonstrates the use of OpenCV and NumPy to perform image histogram analysis and apply histogram equalization for enhancing contrast. The project includes a function to plot the histogram for each color channel (blue, green, red) and a separate function that applies CLAHE (Contrast Limited Adaptive Histogram Equalization) on the luminance channel in the LAB color space. The original and equalized histograms are visualized using Matplotlib, and both the original and processed images are displayed using OpenCV.